Added ThreadLocalTest to "src/tests"
