+++ src/classes/java/lang/ThreadLocal.java
@@ -101,23 +101,6 @@ public class ThreadLocal<T> {
     public static <S> ThreadLocal<S> withInitial(Supplier<? extends S> supplier) {
       return new SuppliedThreadLocal<>(supplier);
     }
-    @Test
-    class threadSafeFormatter{
-      public static ThreadLocal<SimpleDateFormat> df = ThreadLocal.withInitial(() 
-          -> new SimpleDateFormat("yyyy-MM-dd"));
-  public class App  {
-      public static void main(String[] args) throws Exception {
-          threadSafeFormatter tf = new threadSafeFormatter();
-          Thread t1 = new Thread();
-          t1.start();
-      }
-    }
-      public static String birthDate(int userId){
-          Date birthdDate = new Date(userId);
-          final SimpleDateFormat df = threadSafeFormatter.df.get();
-          return df.format(birthdDate);
-      }
-  }
   
     // we need to preserve the modifiers since this might introduce races (supplier could be shared)
     private final Supplier<? extends E> sup;
